refactor: components를 common과 각 page 별로 분리
