+++ src/App.tsx
@@ -1,5 +1,5 @@
 import AppRoutes from "routes/AppRoutes";
-import Layout from "components/common/layout";
+import Layout from "common/components/layout";
 
 //아래 Route외 html Layout 컴포넌트로 변경 예정
 function App() {
