+++ src/common/components/labelTextarea/LabelTextarea.types.ts
@@ -1,0 +1,5 @@
+import { TextareaProps } from "common/components/textarea/Textarea.types";
+
+export type LabelTextareaProps = TextareaProps & {
+  title: string;
+};
