+++ src/common/components/layout/index.tsx
@@ -3,8 +3,8 @@ import { css } from "@emotion/react";
 import { HTMLAttributes, ReactNode } from "react";
 import { Link } from "react-router-dom";
 import colors from "styles/color";
-import Menus from "components/common/menus";
-import { Search } from "components/icons";
+import Menus from "common/components/menus";
+import { Search } from "common/components/icons";
 
 interface Props extends HTMLAttributes<HTMLDivElement> {
   children: ReactNode;
