+++ src/common/components/menus/index.tsx
@@ -1,6 +1,6 @@
 /** @jsxImportSource @emotion/react */
 import { css } from "@emotion/react";
-import { Dashboard, Menu, Order, Review } from "components/icons";
+import { Dashboard, Menu, Order, Review } from "common/components/icons";
 import { ROUTES } from "constants/common";
 import { Link, useLocation } from "react-router-dom";
 import colors from "styles/color";
