--- src/components/common/LabelTextarea/LabelTextarea.types.ts
@@ -1,5 +1,0 @@
-import { TextareaProps } from "components/common/Textarea/Textarea.types";
-
-export type LabelTextareaProps = TextareaProps & {
-  title: string;
-};
