--- src/components/common/index.ts
@@ -1,3 +1,0 @@
-export * from "./Button";
-export * from "./Input";
-export * from "./Textarea";
