--- src/components/index.ts
@@ -1,3 +1,0 @@
-//컴포넌트 모으는곳
-
-export * from "./common";
