--- src/components/menu/index.ts
@@ -1,1 +1,0 @@
-export { default as MenuOption } from "./MenuOption";
