+++ src/pages/menuDetail/components/index.ts
@@ -1,0 +1,1 @@
+export { default as MenuOption } from "./menuOption";
