+++ src/pages/menuDetail/components/menuOption/index.tsx
@@ -1,6 +1,6 @@
 /** @jsxImportSource @emotion/react */
 import { css } from "@emotion/react";
-import { Delete, Edit } from "components/icons";
+import { Delete, Edit } from "common/components/icons";
 import useClickOutside from "hooks/useClickOutside";
 import { HTMLAttributes, useState } from "react";
 import colors from "styles/color";
