+++ src/pages/userDetail/index.tsx
